Changed int -> size_t when needed. Fixed allocated vs len polytics

+++ src/pythia_buf_exports.c
@@ -46,23 +46,23 @@ void check_size(const pythia_buf_t *buf, int size) {
 
 void bn_read_buf(bn_t b, const pythia_buf_t *buf) {
     check_size(buf, 2);
-    bn_read_bin(b, buf->p + 1, buf->len - 1);
+    bn_read_bin(b, buf->p + 1, (int)(buf->len - 1));
     b->sign = buf->p[0];
 }
 
 void gt_read_buf(gt_t g, const pythia_buf_t *buf) {
     check_size(buf, 1);
-    gt_read_bin(g, buf->p, buf->len);
+    gt_read_bin(g, buf->p, (int)buf->len);
 }
 
 void g1_read_buf(g1_t g, const pythia_buf_t *buf) {
     check_size(buf, 1);
-    g1_read_bin(g, buf->p, buf->len);
+    g1_read_bin(g, buf->p, (int)buf->len);
 }
 
 void g2_read_buf(g2_t g, const pythia_buf_t *buf) {
     check_size(buf, 1);
-    g2_read_bin(g, buf->p, buf->len);
+    g2_read_bin(g, buf->p, (int)buf->len);
 }
 
 void bn_write_buf(pythia_buf_t *buf, bn_t b) {
@@ -70,26 +70,26 @@ void bn_write_buf(pythia_buf_t *buf, bn_t b) {
     check_size(buf, size);
     bn_write_bin(buf->p + 1, size - 1, b);
     buf->p[0] = (uint8_t )b->sign;
-    buf->len = size;
+    buf->len = (size_t)size;
 }
 
 void g2_write_buf(pythia_buf_t *buf, g2_t e) {
     int size = g2_size_bin(e, 1);
     check_size(buf, size);
     g2_write_bin(buf->p, size, e, 1);
-    buf->len = size;
+    buf->len = (size_t)size;
 }
 
 void gt_write_buf(pythia_buf_t *buf, gt_t g) {
     int size = gt_size_bin(g, 1);
     check_size(buf, size);
     gt_write_bin(buf->p, size, g, 1);
-    buf->len = size;
+    buf->len = (size_t)size;
 }
 
 void g1_write_buf(pythia_buf_t *buf, g1_t g) {
     int size = g1_size_bin(g, 1);
     check_size(buf, size);
     g1_write_bin(buf->p, size, g, 1);
-    buf->len = size;
+    buf->len = (size_t)size;
 }
